requirements.txt
PyGithub
 
# Tagging tool dependencies
sentence-transformers>=2.2.2
openai>=1.0.0
pyyaml
numpy
tqdm